refactor(session): share restart recovery expectations
